requirements.txt
git+https://github.com/openai/CLIP.git
easydict==1.9
faiss_gpu==1.7.2
ftfy==6.0.3
matplotlib==3.3.3
nltk==3.6.7
numpy==1.19.4
pandas==1.1.5
Pillow==8.4.0
PyYAML==6.0
regex==2022.6.2
scikit_learn==1.1.2
scipy==1.5.4
setuptools==50.3.2
termcolor==1.1.0
tqdm==4.52.0
torch==1.9.0
torchvision==0.10.0
torchaudio==0.9.0 
